+++ backend/bot.js
@@ -1,4 +1,4 @@
-const {sample, sampleSize, pull} = require("lodash");
+const {sample, sampleSize, pull, times} = require("lodash");
 const Player = require("./player");
 const logger = require("./logger");
 
@@ -16,21 +16,21 @@ module.exports = class extends Player {
   }
 
   getPack(pack) {
-    let min = Math.min(this.picksPerPack,pack.length);
-    for (var i = 0; i < min; i++) {
+    const cardsToPick = Math.min(this.picksPerPack, pack.length);
+    times(cardsToPick, () => {
       const randomPick = sample(pack);
       logger.info(`GameID: ${this.gameId}, Bot, picked: ${randomPick.name}`);
       this.picks.push(randomPick.name);
       pull(pack, randomPick);
-    }
+    });
 
     // burn cards
     const cardsToBurn = Math.min(this.burnsPerPack, pack.length);
-    for (const i = 0; i < cardsToBurn; i++) {
-      const randomPick = sampleSize(pack);
+    times(cardsToBurn, () => {
+      const randomPick = sample(pack);
       logger.info(`GameID: ${this.gameId}, Bot, burnt: ${randomPick.name}`);
       pull(pack, randomPick);
-    }
+    });
 
     this.emit("pass", pack);
   }
