+++ backend/game.js
@@ -199,70 +199,7 @@ module.exports = class Game extends Room {
     super.join(sock);
     this.logger.debug(`${sock.name} joined the game`);
 
-    function regularDraftPickDelegate(autopickIndexes, burnPickCardIds) {
-      autopickIndexes.sort(function(a, b){return b-a;});
-      let cards=[];
-      let pack = this.packs.shift();
-      for (var i = 0; i < autopickIndexes.length; i++) {
-        cards.push( pack.splice(autopickIndexes[i], 1)[0]);
-        logger.info(`GameID: ${this.GameId}, player ${this.name}, picked: ${cards[i].name}`);
-        this.draftLog.pack.push( [`--> ${cards[i].name}`].concat(pack.map(x => `    ${x.name}`)) );
-        this.pool.push(cards[i]);
-        let pickcard = cards[i].name;
-        if (cards[i].foil === true)
-          pickcard = "*" + pickcard + "*";
-        this.picks.push(pickcard);
-        this.updateDraftStats(this.draftLog.pack[ this.draftLog.pack.length-autopickIndexes.length ], this.pool);
-        this.send("add", cards[i]);
-      }
-
-      // Remove burned cards from pack
-      remove(pack, (card) => burnPickCardIds.includes(card.cardId));
-
-      let [next] = this.packs;
-      if (!next)
-        this.time = 0;
-      else
-        this.sendPack(next);
-      this.autopickIndex = [];
-      this.emit("pass", pack);
-    }
-
-    function decadentDraftPickDelegate(index) {
-      index.sort(function(a, b){return b-a;});
-      let cards=[];
-
-      const pack = this.packs.shift();
-      for (var i = 0; i < index.length; i++) {
-        cards.push( pack.splice(index[i], 1)[0]);
-        this.draftLog.pack.push( [`--> ${cards[i].name}`].concat(pack.map(x => `    ${x.name}`)) );
-        this.updateDraftStats(this.draftLog.pack[ this.draftLog.pack.length-1 ], this.pool);
-
-        let pickcard = cards[i].name;
-        if (cards[i].foil === true)
-          pickcard = "*" + pickcard + "*";
-
-        this.pool.push(cards[i]);
-        this.picks.push(pickcard);
-        this.send("add", cards[i]);
-      }
-      let [next] = this.packs;
-      if (!next)
-        this.time = 0;
-      else
-        this.sendPack(next);
-
-      // Discard the rest of the cards in the pack
-      // after one is chosen.
-      pack.length = 0;
-
-      this.autopickIndex = [];
-      this.emit("pass", pack);
-    }
-
-    const draftPickDelegate = this.isDecadent ? decadentDraftPickDelegate : regularDraftPickDelegate;
-
-    const h = new Human(sock, draftPickDelegate, this.picksPerPack, this.burnsPerPack, this.id);
+    const h = new Human(sock, this.picksPerPack, this.burnsPerPack, this.id);
     if (h.id === this.hostID) {
       h.isHost = true;
       sock.once("start", this.start.bind(this));
